commit 0316-03 修改完数据格式的问题

+++ CSBP/src/new_node_first_master_problem.cpp
@@ -21,7 +21,7 @@ bool SolveNewNodeFirstMasterProblem(
 
 	// set cons bound
 	int item_types_num = Values.item_types_num;
-	for (size_t k = 0; k < item_types_num; k++)
+	for (int k = 0; k < item_types_num; k++)
 	{
 		int item_type_demand = Lists.all_item_types_list[k].demand;
 		con_min.add(IloNum(item_type_demand)); // cons > demand
@@ -33,15 +33,15 @@ bool SolveNewNodeFirstMasterProblem(
 	Model_MP.add(Cons_MP); // add cons to the model
 
 	// set model matrix 
-	size_t cols_num = this_node.model_matrix.size();
-	size_t rows_num = item_types_num;
+	int cols_num = this_node.model_matrix.size();
+	int rows_num = item_types_num;
 
-	for (size_t col = 0; col < cols_num; col++)
+	for (int col = 0; col < cols_num; col++)
 	{
 		IloNum obj_para = 1;
 		IloNumColumn CplexCol = Obj_MP(obj_para); // Init a col
 
-		for (size_t row = 0; row < rows_num; row++) // set rows in this col
+		for (int row = 0; row < rows_num; row++) // set rows in this col
 		{
 			IloNum row_para = this_node.model_matrix[col][row];
 			CplexCol += Cons_MP[row](row_para); // set coeff
@@ -53,7 +53,7 @@ bool SolveNewNodeFirstMasterProblem(
 		if (col == parent_node.var_to_branch_idx)
 		{
 			IloNum to_branch_val = this_node.var_to_branch_int_val_final;
-			printf("\n	x_var_%zd is set as %f, to be branched", col + 1, to_branch_val);
+			printf("\n	x_var_%d is set as %f, to be branched", col + 1, to_branch_val);
 
 			IloNumVar Var(CplexCol, to_branch_val, to_branch_val, ILOFLOAT, X_name.c_str()); // Init and set var
 			Vars_MP.add(Var);
@@ -63,16 +63,16 @@ bool SolveNewNodeFirstMasterProblem(
 		else 
 		{
 			// Case 2.1: var of this col is NOT a branched - var in previous Nodes	
-			size_t branched_num = parent_node.branched_vars_int_val_list.size();
+			int branched_num = parent_node.branched_vars_int_val_list.size();
 			bool find_flag = 0;
 
-			for (size_t idx = 0; idx < branched_num; idx++) // loop of all branched-vars in previous Nodes
+			for (int idx = 0; idx < branched_num; idx++) // loop of all branched-vars in previous Nodes
 			{
 				int branched_col = parent_node.branched_vars_idx_list[idx];
 				if (col == branched_col) // var of this col is a branched-var in Parent Node
 				{
 					IloNum branched_val = parent_node.branched_vars_int_val_list[idx];
-					printf("\n	x_var_%zd is set as %d, branched", col + 1,  branched_val);
+					printf("\n	x_var_%d is set as %f, branched", col + 1,  branched_val);
 
 					IloNumVar Var(CplexCol, branched_val, branched_val, ILOFLOAT, X_name.c_str()); // Init and set var
 					Vars_MP.add(Var);
@@ -113,7 +113,7 @@ bool SolveNewNodeFirstMasterProblem(
 		printf("\n	Node_%d MP-1 is FEASIBLE\n", this_node.index);
 		printf("\n	OBJ of Node_%d MP-1 is %f\n\n", this_node.index, MP_cplex.getValue(Obj_MP));
 
-		for (size_t col = 0; col < cols_num; col++)
+		for (int col = 0; col < cols_num; col++)
 		{
 			IloNum soln_val = MP_cplex.getValue(Vars_MP[col]);
 			if (soln_val > 0) // feasible soln > 0
@@ -126,42 +126,42 @@ bool SolveNewNodeFirstMasterProblem(
 					if (soln_int_val >= 1)
 					{
 						int_num++;
-						printf("	var_x_%zd = %f int\n", col + 1, soln_val);
+						printf("	var_x_%d = %f int\n", col + 1, soln_val);
 					}
 				}
 				else
 				{
-					printf("	var_x_%zd = %f\n", col + 1, soln_val);
+					printf("	var_x_%d = %f\n", col + 1, soln_val);
 				}
 			}
 		}
 
 
 		printf("\n	BRANCHED VARS: \n\n");
-		size_t branched_num = this_node.branched_vars_int_val_list.size();
-		for (size_t k = 0; k < branched_num; k++)
+		int branched_num = this_node.branched_vars_int_val_list.size();
+		for (int k = 0; k < branched_num; k++)
 		{
-			printf("	var_x_%d = %d branched \n",
+			printf("	var_x_%d = %f branched \n",
 				this_node.branched_vars_idx_list[k] + 1, this_node.branched_vars_int_val_list[k]);
 		}
 
 		printf("\n	DUAL PRICES: \n\n");
 		this_node.dual_prices_list.clear();
 
-		for (size_t row = 0; row < rows_num; row++)
+		for (int row = 0; row < rows_num; row++)
 		{
 			double dual_price = MP_cplex.getDual(Cons_MP[row]);
-			printf("	dual_r_%zd = %f\n", row + 1, dual_price);
+			printf("	dual_r_%d = %f\n", row + 1, dual_price);
 			this_node.dual_prices_list.push_back(dual_price);
 		}
 
 		this_node.lower_bound = MP_cplex.getValue(Obj_MP);
 		printf("\n	Node_%d MP-1:\n", this_node.index);
 		printf("\n	Lower Bound = %f", this_node.lower_bound);
-		printf("\n	NUM of all solns = %zd", cols_num);
+		printf("\n	NUM of all solns = %d", cols_num);
 		printf("\n	NUM of fsb solns = %d", fsb_num);
 		printf("\n	NUM of int solns = %d", int_num);
-		printf("\n	NUM of branched-vars = %zd\n", branched_num);
+		printf("\n	NUM of branched-vars = %d\n", branched_num);
 	}
 
 	MP_cplex.end();
