+++ CSBP/src/update_master_problem.cpp
@@ -23,9 +23,9 @@ bool SolveUpdateMasterProblem(
 
 	// add the new col ro the model of MP 
 	int rows_num = Values.item_types_num;
-	for (size_t row = 0; row < rows_num; row++)
+	for (int row = 0; row < rows_num; row++)
 	{
-		IloNum row_para= this_node.new_col[row];
+		IloNum row_para = this_node.new_col[row];
 		CplexCol += Cons_MP[row](row_para);
 	}
 
@@ -38,7 +38,7 @@ bool SolveUpdateMasterProblem(
 	CplexCol.end(); // end this IloNumColumn object
 
 	// solve the new updated MP
-	printf("\n\n####################### Node_%d MP-%d CPLEX SOLVING START #######################\n\n",this_node.index, this_node.iter + 1);
+	printf("\n\n####################### Node_%d MP-%d CPLEX SOLVING START #######################\n\n", this_node.index, this_node.iter + 1);
 	IloCplex MP_cplex(Env_MP);
 	MP_cplex.extract(Model_MP);
 	//MP_cplex.exportModel("updateMasterProblem.lp");
@@ -49,8 +49,8 @@ bool SolveUpdateMasterProblem(
 	// print fsb-solns of the updated MP
 	int fsb_num = 0;
 	int int_num = 0;
-	size_t cols_num = this_node.model_matrix.size();
-	for (size_t col = 0; col < cols_num; col++)
+	int cols_num = this_node.model_matrix.size();
+	for (int col = 0; col < cols_num; col++)
 	{
 		IloNum soln_val = MP_cplex.getValue(Vars_MP[col]);
 		if (soln_val > 0) // feasible soln > 0
@@ -63,45 +63,45 @@ bool SolveUpdateMasterProblem(
 				if (soln_int_val >= 1)
 				{
 					int_num++;
-					printf("	var_x_%zd = %f int\n", col + 1, soln_val);
+					printf("	var_x_%d = %f int\n", col + 1, soln_val);
 				}
 			}
 			else
 			{
-				printf("	var_x_%zd = %f\n", col + 1, soln_val);
-			}					
+				printf("	var_x_%d = %f\n", col + 1, soln_val);
+			}
 		}
 	}
 
 	printf("\n	BRANCHED VARS: \n\n");
-	size_t branched_num = this_node.branched_vars_int_val_list.size();
-	int var_idx=-1;
-	int var_int_val = -1;
-	for (size_t k = 0; k < branched_num; k++)
+	int branched_num = this_node.branched_vars_int_val_list.size();
+	int var_idx = -1;
+	double var_int_val = -1;
+	for (int k = 0; k < branched_num; k++)
 	{
 		var_idx = this_node.branched_vars_idx_list[k] + 1;
 		var_int_val = this_node.branched_vars_int_val_list[k];
-		printf("	var_x_%d = %d branched \n", var_idx, var_int_val);
+		printf("	var_x_%d = %f branched \n", var_idx, var_int_val);
 	}
 
 	// print and store dual-prices of MP cons
 	this_node.dual_prices_list.clear();
 	printf("\n	DUAL PRICES: \n\n");
 
-	for (size_t row = 0; row < rows_num; row++)
+	for (int row = 0; row < rows_num; row++)
 	{
 		double dual_val = MP_cplex.getDual(Cons_MP[row]); // get dual-prices of all cons
-		printf("	dual_r_%zd = %f\n", row + 1, dual_val);
+		printf("	dual_r_%d = %f\n", row + 1, dual_val);
 		this_node.dual_prices_list.push_back(dual_val);
 	}
 
 	this_node.lower_bound = MP_cplex.getValue(Obj_MP);
-	printf("\n	Node_%d MP-%d:\n", this_node.index,this_node.iter);
+	printf("\n	Node_%d MP-%d:\n", this_node.index, this_node.iter);
 	printf("\n	Lower Bound = %f", this_node.lower_bound);
-	printf("\n	NUM of now solns = %zd", cols_num);
+	printf("\n	NUM of now solns = %d", cols_num);
 	printf("\n	NUM of fsb-solns = %d", fsb_num);
 	printf("\n	NUM of int-solns = %d", int_num);
-	printf("\n	NUM of branched-vars = %zd\n", branched_num);
+	printf("\n	NUM of branched-vars = %d\n", branched_num);
 
 	MP_cplex.end();
 	return MP_flag;
@@ -120,7 +120,7 @@ bool SolveFinalMasterProblem(
 	int item_types_num = Values.item_types_num;
 
 	// solve the final MP of this Node
-	printf("\n\n####################### Node_%d MP-final CPLEX SOLVING START #######################\n\n",this_node.index);
+	printf("\n\n####################### Node_%d MP-final CPLEX SOLVING START #######################\n\n", this_node.index);
 	IloCplex MP_cplex(Model_MP);
 	MP_cplex.extract(Model_MP);
 	//MP_cplex.exportModel("FinalMasterProblem.lp");
@@ -131,8 +131,8 @@ bool SolveFinalMasterProblem(
 
 	this_node.lower_bound = MP_cplex.getValue(Obj_MP);
 
-	size_t cols_num = this_node.model_matrix.size();
-	for (size_t col = 0; col < cols_num; col++)
+	int cols_num = this_node.model_matrix.size();
+	for (int col = 0; col < cols_num; col++)
 	{
 		IloNum soln_val = MP_cplex.getValue(Vars_MP[col]);
 		this_node.all_solns_val_list.push_back(soln_val); // 1. Node all solns (including zero-solns)
@@ -147,12 +147,12 @@ bool SolveFinalMasterProblem(
 					this_node.int_solns_val_list.push_back(soln_val); // 4. Node int-solns
 					this_node.int_solns_idx_list.push_back(col); // 5. Node int-solns' index
 
-					printf("	var_x_%zd = %f int\n", col + 1, soln_val);
+					printf("	var_x_%d = %f int\n", col + 1, soln_val);
 				}
 			}
 			else // non-int-solns
 			{
-				printf("	var_x_%zd = %f\n", col + 1, soln_val);
+				printf("	var_x_%d = %f\n", col + 1, soln_val);
 			}
 
 			this_node.fsb_solns_val_list.push_back(soln_val); // 2. Node feasible (i.e. non-zero-solns) solns 
@@ -162,25 +162,25 @@ bool SolveFinalMasterProblem(
 
 
 	printf("\n	BRANCHED VARS: \n\n");
-	size_t branched_num = this_node.branched_vars_int_val_list.size();
+	int branched_num = this_node.branched_vars_int_val_list.size();
 	int var_idx = -1;
-	int var_int_val = -1;
-	for (size_t k = 0; k < branched_num; k++)
+	double var_int_val = -1;
+	for (int k = 0; k < branched_num; k++)
 	{
 		var_idx = this_node.branched_vars_idx_list[k] + 1;
 		var_int_val = this_node.branched_vars_int_val_list[k];
-		printf("	var_x_%d = %d branched \n", var_idx, var_int_val);
+		printf("	var_x_%d = %f branched \n", var_idx, var_int_val);
 	}
 
-	size_t fsb_num= this_node.fsb_solns_val_list.size();
-	size_t int_num = this_node.int_solns_idx_list.size();
+	int fsb_num = this_node.fsb_solns_val_list.size();
+	int int_num = this_node.int_solns_idx_list.size();
 	this_node.lower_bound = MP_cplex.getValue(Obj_MP);
 	printf("\n	Node_%d MP-final:\n", this_node.index);
 	printf("\n	Lower Bound = %f", this_node.lower_bound);
-	printf("\n	NUM of all solns = %zd", cols_num);
-	printf("\n	NUM of fsb-solns = %zd", fsb_num);
-	printf("\n	NUM of int-solns = %zd", int_num);
-	printf("\n	NUM of branched-vars = %zd\n", branched_num);
+	printf("\n	NUM of all solns = %d", cols_num);
+	printf("\n	NUM of fsb-solns = %d", fsb_num);
+	printf("\n	NUM of int-solns = %d", int_num);
+	printf("\n	NUM of branched-vars = %d\n", branched_num);
 
 	MP_cplex.end();
 	return MP_flag;
